move p2p handler to p2pcontroller
